+++ merchantops-api/src/main/java/com/renda/merchantops/api/controller/DevController.java
@@ -5,6 +5,7 @@ import com.renda.merchantops.common.exception.BizException;
 import com.renda.merchantops.common.exception.ErrorCode;
 import com.renda.merchantops.common.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,6 +13,7 @@ import java.util.Map;
 
 @RestController
 @RequestMapping("/api/v1/dev")
+@Tag(name = "Dev Test")
 public class DevController {
 
     @Operation(summary = "Ping dev endpoint")